Unifed all race monitor API communications into one core generic function that can be passed URL snippets and post data dictionaries in order to pull different information. This reduced the code repetiveness in core.py greatly.
The generic function will raise an exception if the query to race monitor was not successful. The error handling needs to be fleshed out more to cover various HTTP errors the Race Monitor API docs list as needing to be handled.

Added localization support

Moved the API token filepath and constant to const.py This is still a development-only system. If/when this software is released commercially the issued api token will need to be included in the application and encrypted.